Changed ProductSKU to datatype "long", and reader in search and products

+++ InventorySystem/Models/ProductsGET.cs
@@ -4,14 +4,14 @@ namespace InventorySystem.Models
     public struct ProductsGET
     {
         public int ProductID { get; set; }
-        public int ProductSKU { get; set; }
+        public long ProductSKU { get; set; }
         public string ProductName { get; set; }
         public int ProductPrice { get; set; }
         public int ProductVariableCost { get; set; }
         public string ProductBrand { get; set; }
         //public int ProductMaterialsOrderID { get; set; }
         public ProductsGET(int id,
-                       int sku,
+                       long sku,
                        string name,
                        int price,
                        int variablecost,
